Crate rustc_traits

New recursive solver modeled on Chalk’s recursive solver. Most of the guts are broken up into modules; see the comments in those modules.

Modules

chalk 🔒

Calls chalk-solve to solve a ty::Predicate

codegen 🔒

dropck_outlives 🔒

evaluate_obligation 🔒

implied_outlives_bounds 🔒

Provider for the implied_outlives_bounds query. Do not call this query directory. See rustc_trait_selection::traits::query::type_op::implied_outlives_bounds.

normalize_erasing_regions 🔒

normalize_projection_ty 🔒

type_op 🔒

Functions

provide

type_op_ascribe_user_type_with_span

The core of the type_op_ascribe_user_type query: for diagnostics purposes in NLL HRTB errors, this query can be re-run to better track the span of the obligation cause, and improve the error message. Do not call directly unless you’re in that very specific context.

type_op_prove_predicate_with_cause

The core of the type_op_prove_predicate query: for diagnostics purposes in NLL HRTB errors, this query can be re-run to better track the span of the obligation cause, and improve the error message. Do not call directly unless you’re in that very specific context.