Module rustc_traits::implied_outlives_bounds

Provider for the implied_outlives_bounds query. Do not call this query directory. See rustc_trait_selection::traits::query::type_op::implied_outlives_bounds.

Functions

compute_implied_outlives_bounds 🔒

implied_bounds_from_components 🔒

When we have an implied bound that T: 'a, we can further break this down to determine what relationships would have to hold for T: 'a to hold. We get to assume that the caller has validated those relationships.

implied_outlives_bounds 🔒

provide 🔒