Function rustc_traits::implied_outlives_bounds::implied_bounds_from_components

fn implied_bounds_from_components<'tcx>(
    sub_region: Region<'tcx>,
    sup_components: SmallVec<[Component<'tcx>; 4]>
) -> Vec<OutlivesBound<'tcx>>

When we have an implied bound that T: 'a, we can further break this down to determine what relationships would have to hold for T: 'a to hold. We get to assume that the caller has validated those relationships.