Module rustc_traits::type_op

Structs

AscribeUserTypeCx 🔒

Functions

provide 🔒

type_op_ascribe_user_type 🔒

type_op_ascribe_user_type_with_span

The core of the type_op_ascribe_user_type query: for diagnostics purposes in NLL HRTB errors, this query can be re-run to better track the span of the obligation cause, and improve the error message. Do not call directly unless you’re in that very specific context.

type_op_eq 🔒

type_op_normalize 🔒

type_op_normalize_fn_sig 🔒

type_op_normalize_poly_fn_sig 🔒

type_op_normalize_predicate 🔒

type_op_normalize_ty 🔒

type_op_prove_predicate 🔒

type_op_prove_predicate_with_cause

The core of the type_op_prove_predicate query: for diagnostics purposes in NLL HRTB errors, this query can be re-run to better track the span of the obligation cause, and improve the error message. Do not call directly unless you’re in that very specific context.

type_op_subtype 🔒